small fixes

+++ src/Controller/AdminConfigController.php
@@ -2,13 +2,13 @@
 
 namespace MartenaSoft\Menu\Controller;
 
+use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use MartenaSoft\Common\Controller\AbstractAdminBaseController;
 use MartenaSoft\Common\Library\CommonValues;
 use MartenaSoft\Menu\Entity\Config;
 use MartenaSoft\Menu\Entity\ConfigSearch;
-use MartenaSoft\Menu\Entity\Menu;
 use MartenaSoft\Menu\Form\ConfigType;
 use MartenaSoft\Menu\Form\SearchFormType;
 use MartenaSoft\Menu\Repository\ConfigRepository;
@@ -20,19 +20,25 @@ use Symfony\Component\HttpFoundation\Response;
 class AdminConfigController extends AbstractAdminBaseController
 {
     public const CONFIG_SAVED_SUCCESS_MESSAGE = 'Config saved success';
+    public const CONFIG_DELETE_ERROR_DEFAULT_MESSAGE = "You can't delete default configure.";
+    public const CONFIG_DELETE_ERROR_FOREIGN_MESSAGE = "This configurations used in menus";
+
 
     private EntityManagerInterface $entityManager;
     private LoggerInterface $logger;
     private ConfigRepository $configRepository;
+    private MenuRepository $menuRepository;
 
     public function __construct(
         EntityManagerInterface $entityManager, 
         LoggerInterface $logger, 
-        ConfigRepository $configRepository
+        ConfigRepository $configRepository,
+        MenuRepository $menuRepository
     ) {
         $this->entityManager = $entityManager;
         $this->logger = $logger;
         $this->configRepository = $configRepository;
+        $this->menuRepository = $menuRepository;
     }
 
     public function index(Request $request, PaginatorInterface $paginator, int $page = 1): Response
@@ -61,6 +67,8 @@ class AdminConfigController extends AbstractAdminBaseController
     {
         if (empty($id)) {
             $menuConfigEntity = new Config();
+            $count = $this->configRepository->count(["isDefault" => true]);
+            $menuConfigEntity->setIsDefault($count == 0);
             $this->entityManager->persist($menuConfigEntity);
         } else {
             $menuConfigEntity = $this->configRepository->find($id);
@@ -71,9 +79,8 @@ class AdminConfigController extends AbstractAdminBaseController
 
         if ($form->isSubmitted() && $form->isValid()) {
             try {
-
                 if ($form->get('isDefault')) {
-                    $updated = $this->configRepository->setAllIsDefaultAsFalse();
+                    $this->configRepository->setAllIsDefaultAsFalse();
                     $menuConfigEntity->setIsDefault(true);
                 }
 
@@ -97,7 +104,7 @@ class AdminConfigController extends AbstractAdminBaseController
         ]);
     }
 
-    public function show(int $id = 0): Response
+    public function view(int $id = 0): Response
     {
         $configEntity = $this->configRepository->find($id);
         if (empty($configEntity)) {
@@ -105,11 +112,22 @@ class AdminConfigController extends AbstractAdminBaseController
                             CommonValues::ERROR_ENTITY_RECORD_NOT_FOUND);
             return $this->redirectToRoute('menu_admin_config_index');
         }
-        return $this->render('@MartenaSoftMenu/admin_config/show.html.twig', [
-            'configEntity' => $configEntity
+
+        $allRoots = $this
+            ->menuRepository
+            ->getAllRootsQueryBuilder()
+            ->andWhere("m.config=:config")
+            ->setParameter("config", $configEntity)
+            ->getQuery()
+            ->getResult()
+        ;
+
+        return $this->render('@MartenaSoftMenu/admin_config/view.html.twig', [
+            'configEntity' => $configEntity,
+            'allRoots' => $allRoots
         ]);
     }
-    
+
     public function delete(int $id = 0): Response
     {
         $entity = $this->configRepository->find($id);
@@ -119,9 +137,18 @@ class AdminConfigController extends AbstractAdminBaseController
             return $this->redirectToRoute('menu_admin_config_index');
         }
 
+        if ($entity->isDefault()) {
+            $this->addFlash(CommonValues::FLASH_ERROR_TYPE,
+                            self::CONFIG_DELETE_ERROR_DEFAULT_MESSAGE);
+            return $this->redirectToRoute('menu_admin_config_index');
+        }
+
         try {
             $this->entityManager->remove($entity);
             $this->entityManager->flush();
+        } catch (ForeignKeyConstraintViolationException $exception) {
+            $this->addFlash(CommonValues::FLASH_ERROR_TYPE,
+                            self::CONFIG_DELETE_ERROR_FOREIGN_MESSAGE);
         } catch (\Throwable $exception) {
             $this->logger->error(CommonValues::ERROR_FORM_SAVE_LOGGER_MESSAGE, [
                 'class' => __CLASS__,
